IMPALA-10927: Deflaky TestFetchAndSpooling.test_rows_sent_counters
IMPALA-8957 fixed the flakiness for test by adding a delay via
DEBUG_ACTION BPRS_BEFORE_ADD_ROWS in BlockingPlanRootSink::Send().
BPRS_BEFORE_ADD_ROWS should be used for spool_query_results as false,
BPRS_BEFORE_ADD_BATCH should be used for spool_query_results as true.

Extended dimension is added for TestFetchAndSpooling with
spool_query_results as true. Now the default value of pool_query_results
is set as true so spool_query_results as false is not tested in the test
case.
Since the test uses BPRS_BEFORE_ADD_ROWS if spool_query_results is
not specified as true, it makes DEBUG_ACTION BPRS_BEFORE_ADD_ROWS to be
used for spool_query_results as true. This causes the test flaky since
no delay in BufferedPlanRootSink::Send().
Another issue is that the test uses bool() to convert string to boolean
value, but the function returns true for any non empty string.

This patch changed the extended dimension setting for
spool_query_results as false, and made the test to use the right
DEBUG_ACTION for spool_query_results as true and false.
Also reverted the previous fixing which disabled the test for S3
testing environment.

Testing:
  - Ran the test more than 300 times on local machine.

Change-Id: I790bbe1072357caf8ee11bb37644cf29dc8bea0f