IMPALA-12375: Make DataSource Object persistent
DataSource objects are saved in-memory cache in Catalog server. They are
not persisted to the HMS. The objects are lost after Catalog server is
restarted.
This patch makes DataSource Object persistent by saving DataSource
objects as DataConnector objects with type "impalaDataSource" in HMS.
Since HMS events for DataConnector are not handled, Catalog server
has to refresh DataSource objects when the catalogd becomes active.
Note that this feature is not supported for Apache Hive 3.1 or older
version.

Testing:
 - Added end-to-end test cases with restarting of Catalog server,
   and catalogd HA failover.
 - Passed all-build-options-ub2004.
 - Passed core test.

Change-Id: I500a99142bb62ce873e693d573064ad4ffa153ab