IMPALA-13134 (part 2): Fix flaky test_catalogd_failover_with_sync_ddl
The test_catalogd_failover_with_sync_ddl test which was added to
custom_cluster/test_catalogd_ha.py in previous patch failed on s3.
The test relies on specific timing with a sleep injected via a
debug action so that the DDL query is still running when catalogd
failover is triggered. The failures were caused by slowly restarting
for catalogd on s3.

This patch fixed the issue by increasing the sleep time for s3 builds
and slow builds.

Testing:
 - Ran the test 200 times in a loop on s3.

Change-Id: I15bb6aae23a2f544067f993533e322969372ebd5